Nearly complete with the UAC tab interactions. Just need to properly archive a user and their permissions when the delete button is hit.
Also need to build the "Previous Users" dialog box and the connection
between it and it's respective points with the UAC Tab in the main
widget. What I mean by that is when delete is hit and confirmed, the
user's name and permissions get stored and then displayed in the
Previous Users. Then if a previous user is restored, the restoration
process is prepared properly. If confirmed, then the user and their
permissions are returned to the table widget in the UAC tab.

I did get the calendar buttons in the Reports Tab to display calendar
icons.

I managed to get the Add A User process fully complete. It provides a
good walkthrough for the user to enter in a user correctly and
successfully, along with their permissions.

I created a resource folder for the calendar icon and any other images
that may be needed.

Once I get the mapping of the user permissions from Brian, I'll have a
better understanding of the connection that need to be made for tailored
user profiles. The range goes from a user having minimal access and
permissions all the way up to an admin. The purpose here is to make sure
the application allows interactions based on user permissions, which is
connected to a user's profile and their login.

I could go into greater detail here on what more needs to get done with
each subcomponent, component, and subsystem, but I'll cut off here for
now.